refactor(dispatch): extract envelope_types.py leaf module (PR-1 monolith split)
Moves EnvelopeSpec, EnvelopeResult, EnvelopeGovernError and _AdapterResult
out of dispatch_envelope.py into a new leaf module with zero sibling
imports. dispatch_envelope.py re-exports all four via runtime `from
envelope_types import X` (bare-name form) so patch-string sites and
get_type_hints keep resolving against the facade's globals, and existing
consumer imports (dispatch_cli.py, provider_dispatch.py) are unchanged.

Regenerates the PR-0 census fixture: the coupling count is unchanged
(116), only the discovered module family grows to include envelope_types.

Dispatch-ID: 20260803-211041-pr1-envelope-types
